refactor: give the dump vocabulary a file of its own
save.go held three unrelated things: the value types a dump is described in,
the auto-save connector, and the machinery that writes a table back over the
file it came from. OutputFormat, CompressionType and DumpOptions have no
behaviour beyond reading themselves back and are what DumpDatabase takes;
they lived there only because that is where the first of them was written.

save.go drops from 934 lines to 706, and the twelve tests that pin those
values move with them. The test function count over the package is unchanged.
